requirements.txt
django
djangorestframework
pillow
tensorflow
pymongo
cloudinary
